Move e2e tests to balancer/rls
